+++ src/Polarion/Client/IPolarionClient.cs
@@ -3,7 +3,7 @@ namespace Polarion;
 public interface IPolarionClient
 {
     [RequiresUnreferencedCode("Uses WCF services which require reflection")]
-    Task<Result<WorkItem>> GetWorkItemByIdAsync(string workItemId);
+    Task<Result<WorkItem>> GetWorkItemByIdAsync(string workItemId, string? revision = null);
 
     [RequiresUnreferencedCode("Uses WCF services which require reflection")]
     Task<Result<WorkItem[]>> SearchWorkitemAsync(string query, string order, List<string> field_list);
@@ -51,7 +51,7 @@ public interface IPolarionClient
     Task<Result<string[]>> GetRevisionsIdsByWorkItemIdAsync(string workItemId);
 
     [RequiresUnreferencedCode("Uses WCF services which require reflection")]
-    Task<Result<WorkItem[]>> GetWorkItemRevisionsByIdAsync(string workItemId, int maxRevisions = -1);
+    Task<Result<Dictionary<string, WorkItem>>> GetWorkItemRevisionsByIdAsync(string workItemId, int maxRevisions = -1);
 
     [RequiresUnreferencedCode("Uses WCF services which require reflection")]
     Task<Result<Module[]>> GetModuleRevisionsByLocationAsync(string location, int maxRevisions = -1);